+++ Source/mesquite/trees/ArcTree/ArcTree.java
@@ -10,11 +10,28 @@ Mesquite is distributed under the terms of the GNU Lesser General Public License
  */
 package mesquite.trees.ArcTree;
 
-import java.util.*;
-import java.awt.*;
+import java.awt.BasicStroke;
+import java.awt.Color;
+import java.awt.Graphics;
+import java.awt.geom.Path2D;
+import java.awt.geom.Rectangle2D;
+import java.util.Enumeration;
+import java.util.Vector;
 
-import mesquite.lib.*;
-import mesquite.lib.duties.*;
+import mesquite.lib.CommandChecker;
+import mesquite.lib.EmployeeNeed;
+import mesquite.lib.MesquiteCommand;
+import mesquite.lib.MesquiteDouble;
+import mesquite.lib.MesquiteFile;
+import mesquite.lib.MesquiteInteger;
+import mesquite.lib.MesquiteModule;
+import mesquite.lib.MesquiteNumber;
+import mesquite.lib.MesquiteString;
+import mesquite.lib.MesquiteThread;
+import mesquite.lib.MesquiteTrunk;
+import mesquite.lib.Snapshot;
+import mesquite.lib.duties.DrawTree;
+import mesquite.lib.duties.NodeLocsVH;
 import mesquite.lib.tree.DiagonalRootDrawer;
 import mesquite.lib.tree.MesquiteTree;
 import mesquite.lib.tree.SquareTipDrawer;
@@ -24,9 +41,7 @@ import mesquite.lib.tree.TreeDrawing;
 import mesquite.lib.ui.ColorDistribution;
 import mesquite.lib.ui.GraphicsUtil;
 import mesquite.lib.ui.MesquiteSubmenuSpec;
-import mesquite.trees.lib.*;
-
-import java.awt.geom.*;
+import mesquite.trees.lib.DrawTreeUtil;
 /* ======================================================================== */
 public class ArcTree extends DrawTree implements DiagonalRootDrawer, SquareTipDrawer {
 	public String getName() {
@@ -226,7 +241,7 @@ class ArcTreeDrawing extends TreeDrawing  {
 		calculateLines(tree, drawnRoot);
 		edgewidth = preferredEdgeWidth;
 		if (treeDisplay.getTaxonSpacing()<edgewidth+2) {
-			edgewidth= treeDisplay.getTaxonSpacing()-2;
+			edgewidth= (int)treeDisplay.getTaxonSpacing()-2;
 			if (edgewidth<2)
 				edgewidth=2;
 		}
